+++ PartyFund.Presentation.UI.Common/ViewModels/DepositDetailViewModel.cs
@@ -10,7 +10,7 @@ namespace PartyFund.Presentation.UI.Common.ViewModels
     {
         public int ID { get; set; }
         public int UserID { get; set; }
-        public Nullable<int> TotalSpendMoney { get; set; }
+        public Nullable<decimal> TotalSpendMoney { get; set; }
         public Nullable<System.DateTime> DepositDate { get; set; }
         public string CreatedBy { get; set; }
         public Nullable<System.DateTime> DateCreated { get; set; }
